requirements.txt
yfinance
kafka-python
pyspark
redis
fastapi
uvicorn
cassandra-driver
requests
mysqlclient
psycopg2